refactor: 스켈레톤 폴더 구조 개선 및 재사용 컴포넌트 추출 => what
- A: 평탄한 구조(9개 파일 혼재), Profile/TotalBar 로직이 OtherFeedSkeleton에만 결합
- B: base/feed 도메인 분리 + ProfileSkeleton/TotalBarSkeleton 독립 추출 → 조합형 설계
- 재사용성: Profile/TotalBar 스켈레톤 독립 사용 가능, import 경로 간소화 (개별 경로 → '@/shared/ui/Skeleton')
