+++ src/pages/feed/Feed.tsx
@@ -4,8 +4,7 @@ import TabBar from '../../components/feed/TabBar';
 import MyFeed from '../../components/feed/MyFeed';
 import TotalFeed from '../../components/feed/TotalFeed';
 import MainHeader from '@/components/common/MainHeader';
-import FeedPostSkeleton from '@/shared/ui/Skeleton/FeedPostSkeleton';
-import OtherFeedSkeleton from '@/shared/ui/Skeleton/OtherFeedSkeleton';
+import { FeedPostSkeleton, OtherFeedSkeleton } from '@/shared/ui/Skeleton';
 import writefab from '../../assets/common/writefab.svg';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { getTotalFeeds } from '@/api/feeds/getTotalFeed';
