+++ src/shared/ui/Skeleton/base/Skeleton.styled.ts
