+++ src/shared/ui/Skeleton/base/Skeleton.tsx
