+++ src/shared/ui/Skeleton/base/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Skeleton';
