+++ src/shared/ui/Skeleton/feed/FeedPostSkeleton.styled.ts
