+++ src/shared/ui/Skeleton/feed/ProfileSkeleton.styled.ts
@@ -1,11 +1,5 @@
 import styled from '@emotion/styled';
 
-export const SkeletonContainer = styled.div<{ paddingTop?: number }>`
-  padding-top: ${({ paddingTop }) => (paddingTop !== undefined ? `${paddingTop}px` : '56px')};
-  background-color: var(--color-black-main);
-  min-height: 100vh;
-`;
-
 export const ProfileContainer = styled.div`
   display: flex;
   flex-direction: column;
@@ -46,23 +40,4 @@ export const FollowerAvatars = styled.div`
   flex-direction: row;
   align-items: center;
   gap: 4px;
-`;
-
-export const TotalBarContainer = styled.div`
-  width: 100%;
-  padding: 0 20px;
-  margin-bottom: 4px;
-`;
-
-export const PostContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: 40px 20px;
-  gap: 12px;
-`;
-
-export const PostFooter = styled.div`
-  display: flex;
-  gap: 16px;
-  margin-top: 4px;
 `;
