+++ src/shared/ui/Skeleton/feed/TotalBarSkeleton.styled.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const TotalBarContainer = styled.div`
+  width: 100%;
+  padding: 0 20px;
+  margin-bottom: 4px;
+`;
