+++ src/shared/ui/Skeleton/feed/TotalBarSkeleton.tsx
@@ -1,0 +1,12 @@
+import Skeleton from '../base/Skeleton';
+import { TotalBarContainer } from './TotalBarSkeleton.styled';
+
+const TotalBarSkeleton = () => {
+  return (
+    <TotalBarContainer>
+      <Skeleton.Text width={60} height={14} />
+    </TotalBarContainer>
+  );
+};
+
+export default TotalBarSkeleton;
