+++ src/shared/ui/Skeleton/feed/index.ts
@@ -1,0 +1,4 @@
+export { default as FeedPostSkeleton } from './FeedPostSkeleton';
+export { default as OtherFeedSkeleton } from './OtherFeedSkeleton';
+export { default as ProfileSkeleton } from './ProfileSkeleton';
+export { default as TotalBarSkeleton } from './TotalBarSkeleton';
