+++ src/shared/ui/Skeleton/index.ts
@@ -1,1 +1,6 @@
-export { default } from './Skeleton';
+// Base skeleton
+export { default } from './base/Skeleton';
+
+// Feed skeletons
+export { FeedPostSkeleton, OtherFeedSkeleton, ProfileSkeleton, TotalBarSkeleton } from './feed';
+
